docs(steps): update doc and demo
