+++ src/steps/demos/vertical.vue
@@ -1,27 +1,15 @@
 <template>
-  <div class="cell-base">
-    <t-steps v-model="current" layout="vertical" class="demo-steps">
-      <t-step title="已完成步骤" content="辅助信息文字最多两行"></t-step>
-      <t-step title="当前步骤" content="辅助信息文字最多两行"></t-step>
-      <t-step title="未完成步骤" content="辅助信息文字最多两行"></t-step>
+  <div class="step-demo">
+    <t-steps v-model="current" theme="dot" layout="vertical">
+      <t-step title="事件描述"></t-step>
+      <t-step title="事件描述"></t-step>
+      <t-step title="事件描述"></t-step>
     </t-steps>
   </div>
 </template>
 
-<script lang="ts">
-import { defineComponent, ref } from 'vue';
+<script setup lang="ts">
+import { ref } from 'vue';
 
-export default defineComponent({
-  setup() {
-    const current = ref(0);
-    return {
-      current,
-    };
-  },
-});
+const current = ref(1);
 </script>
-<style lang="less" scoped>
-.demo-steps {
-  padding-top: 12px;
-}
-</style>
